bot/requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
alembic==1.13.3
asyncpg==0.31.0
google-genai==1.67.0
Jinja2==3.1.6
python-docx==1.2.0
python-dotenv==1.0.1
python-pptx==1.0.2